flashcp: Do not export already export gpio pin
There is no requirement to export an already exported gpio pin,
For instance, during a case where the gpio pin was exported manually,
then there will be an instance of that gpio pin already inside sysfs.

We can check if that instance exists, then that pin is considered to be
already exported. In such case, exit quiety from the function.

Change-Id: I73d89216e2c0e169d244120f590095a281a1a48c
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>