flashcp: Do not uexport already unexported gpio pin
There is also no requirement to unexport an already unexported gpio pin,
For instance, during cases where the gpio pin was never exported or it was
unexported manually, Then there will be no instance of that gpio pin inside
sysfs.

We can check if that instance does not exists, then the pin is considered to be
already unexported. In such case, exit quiety from the function.

Change-Id: I065e1cbce7f2e8525a17bf21c6f98f4aa5b23768
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>